pyproject.toml
[tool.poetry]
name = "trygraphql-v1"
version = "0.1.0"
description = ""
authors = ["kofi-teddy"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
graphene-django = "^2.15.0"
django-graphql-jwt = "^0.3.4"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
